requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0

# Database and ORM
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Redis for caching and sessions
redis[hiredis]==5.0.1

# Authentication and security
bcrypt==4.1.2
pyjwt==2.8.0
cryptography>=42.0.0

# HTTP client for external service communication
httpx==0.25.2

# Environment and configuration
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Production WSGI server (optional)
gunicorn==21.2.0

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2